audio/STT-main/requirements.txt
# STT-TTS Integration Requirements
whisper-openai>=20231117
pyaudio>=0.2.11
numpy>=1.21.0
pathlib>=1.0.1

# Optional (for better audio handling)
pydub>=0.25.1
scipy>=1.7.0

# Installation notes:
# 1. Install PyAudio: pip install pyaudio
#    On Windows, you might need: pip install pipwin && pipwin install pyaudio
# 
# 2. Install Whisper: pip install -U openai-whisper
#
# 3. Make sure FFmpeg is installed and in PATH
#
# 4. Ensure Piper TTS is set up in your specified directory